+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Controllers/Upload/ChunkUploadController.cs
@@ -15,9 +15,9 @@ namespace Kendo.Mvc.Examples.Controllers
             public string UploadUid { get; set; }
             public string FileName { get; set; }
             public string ContentType { get; set; }
-            public int ChunkIndex { get; set; }
-            public int TotalChunks { get; set; }
-            public int TotalFileSize { get; set; }
+            public long ChunkIndex { get; set; }
+            public long TotalChunks { get; set; }
+            public long TotalFileSize { get; set; }
         }
 
         public class FileResult
